Ensure signature classes are loaded to allow inlining
In #6280 we found that JIT-generated code for Ruby methods was not
inlining, due to a reported error of "unloaded signature classes".
This error means that not all classes in the target method's
signature were fully loaded and resolved at the point of inlining,
and since the Hotspot JIT cannot directly load classes it would
reject these methods. Specifically for this case, at least the
String argument (called method name) and possibly the RubyModule
argument (called-as class) were flagged as "unloaded", likely
because the simple code body in #6280 never actually accessed
those arguments.

This patch forces the generated code to forcibly load and resolve
all classes used in the jitted method signatures, which should
guarantee they do not appear "unloaded" at native JIT inlining
time. Ideally this will be a fairly cheap operation, performed
only once at classloading time, since the classes in question are
clearly already loaded in the parent classloaders.

For JRuby 9.2, this is only done for the "JVM7" logic, since the
"JVM6" logic typically does not make methods inlinable anyway.