Restore limited shading for jruby-core jar
I think we were confused by lib/jruby.jar when discussing the
jruby-core module and the libraries it shades.
The lib/jruby.jar file for the JRuby distribution shades in all
Java depdendencies, while the actual published jruby-core jar only
shades ASM (hidden), jitescript and jnr-ffi (because they depend
on ASM).

On the new "base" branch, we originally planned for "base" to be
the fully-unshaded jar and "core" to be "base plus shaded Java
dependencies", but this does not match the historical "core" jar
we have been publishing.

This commit restores the limited shading for the jruby-core jar.

The jruby-core module may now be the least useful; it does not
shade in all dependencies it needs, but at the same time it shades
enough (jnr-ffi) to cause version conflicts with other libraries.
I believe we should consider jruby-core deprecated for 9.3 in
favor of jruby-base as the baseline artifact, with core continuing
to be base + relocated ASM and complete being base + all Java deps
and the Ruby stdlib.