Trigger core build and include only complete jar
The core build does not appear to trigger when building jruby-jars
now after the move to the "base" jar setup. I add jruby-core as
a compile dependency here to ensure it builds, and then explicitly
add only the "core complete" jar that we copy from the JRuby dist
lib dir.