Allow ulong range values to coerce to long bits
This adds ulong versions of the numeric-to-long conversion methods
that allow ulong-ranged values to be coerced to a Java long.
Callers must take care to use the resulting long correctly, since
Java will always see it as signed.

Added primarily to support ulong use cases in FFI, and fixes #6376
and a test commented out due to jnr/jffi#87.