Make sure locks are added immediately after acquisition
There's a race doing the lock add after the executeTask, since a
thread event could interrupt the flow and bypass that logic. A
finally would not be appropriate, since an interrupted attempt to
acquire the lock would then stil add it. Instead we just move the
lock add immediately after lockInterruptibly. If we reach this
point, the lock is locked by the current thread, and must be
added.

Fixes #6405 and maybe #6326.