Update enxio to omit poll fd from key list
This long-standing bug allows a null key to get into the list of
keys when using the `poll` selector. One of the elements in the
internal array of keys is expected to be null to reserve a spot
for an additional poll file descriptor, but null elements should
never be returned in the keys list.