Use normalized URI for glob
The createResource call here normalizes a URI like "classpath:" to
our preferred "uri:classloader:" but only uses that normalized
version to check the existence of the resource in question. If the
resource exists, it proceeds to use the un-normalized version in
glob, and fails since "classpath:" is not supported by our glob
internals.

This patch uses the newly normalized path when calling glob
internals, so "classpath:" never leaks into that code.

Fixes #6421