Do not optimize case with null value to switch
Optimized Fixnum switch is only possible when we have a `case`
value and the `when` values are all int-ranged literal numbers.
The NPE in #6440 was due to us also trying to optimize homogeneous
int case/when even if the case value was null (omitted in source).

This patch limits the optimized switch generation to situations
where we have a non-null `case` value.

Fixes #6440.