Simplify detached process thread logic
In #6466 we see that the thread from a Process::detach is raising
an error when the waitpid fails with an error. This does not
appear to match the logic in MRI and also breaks if there is a
"raise" hook installed in the system.

This change should better match MRI logic for the detach thread
and avoid the hook issue reported in #6466. It does not address a
separate bug in the trace hook logic when no Ruby frames are
available.