Filled out to the point I believe RubyClass can ask for the interpreter context via IRMethod.builtInterperterContextForJavaConstructor().  If this is and ExitableInterpreterEngine then we can split execute the method.
At that point we need to ask for ic.getEngineState() and then call
ExitableInterpreterEngine.interpret().  That method will execute up to the
super instr and return a RubyArray of the arguments to the super call.  At
this point we should convert those values for Java CTOR and execute it.  Once
finished we will then just call interpret() again.  The interpreter will then
continue executing at one instr past the original super instr.

So the above description still needs to exist in RubyClass vs the AST stuff
it is doing now.

There is another small problem that this is a startup interpreter of sorts
and this method will still try and promote to being a FULL/JIT method.  We
need to either disable that or possibly automatically convert this to be a
full interpreter (but still disable JIT compilation).