Add depth logic to IR super splitting.
This looks for jump targets and labels.  A label is where we will jump
to.  jumptargets are places which will move to a label.  If we run into
a jump target we need to just notice whether the jump will move forward
past the super call or jump backwards to before the super call.  All
the other jumps should be fine.

From a purely what could we support this is a bit restrictive.  There
are cases where we could support all paths of an if/else where both
branches are guaranteed to call super.  This should be allowed but this
analysis requires flow analysis and the current version is too simple to
support it.  Future work can add this.