Make bundler a proper default gem
We have been versioning bundler directly in stdlib, since that is
how we ship RubyGems and the upgrade process for RG also copies
the bundler files into stdlib. However, since it does not get
installed using the default gem process, RG does not know to look
for an upgraded gem version leading to the issue reported in
issue #6485.

This PR removes the versioned files from stdlib and adds bundler
as a default gem, fixing the reported example.

Fixes #6485 and likely others