Check the retval of waitpid before errno
As described in #6477 it appears that a successful waitpid under
some circumstances is leaving errno set, either to 11 (EAGAIN) or
in some situation 2 (ENOENT). The reasons for this unsanitary
errno handling are not clear, but the logic in JRuby should at
least be checking for a -1 retval before assuming there was an
error. This commit makes that modification which should avoid the
error reported in #6477.

This constitutes a "fix" of sorts, but we still don't know why the
cicumstances in #6477 lead to this errno result.

Fixes #6477.