Disable rlimit features on Windows
We attempted to fake getrlimit behavior on Windows, but CRuby just
leaves it and the related constants (plus setrlimit) unimplemented
when the host platform does not support them. Our fake behavior
led to rubygems/rubygems#4129 which affects all versions of JRuby
up through 9.2.14.0.

This patch markes the related rlimit methods as not implemented
and skips defining the related constants when on Windows.

Fixes #6491