Force time to UTC before getting ms
This method should extract the equivalent ms from the time without
considering timezone, so the absolute value can be subtracted from
a time with a timezone to get the zone offset. Joda normalizes
getMillis to universal time, causing the resulting offset to be
zero.

The patch here forces the Joda DateTime to UTC without altering
the other time fields, allowing us to get equivalent UTC millis
for the given time.

Fixes #6488