Eager load RubyGems to work around #6506
Part of IRB attempts to lazily load RubyGems for its internal find
file logic used by its require and load. Howver the way it
accesses the Gem module does not appear to trigger autoloading,
causing a method it needs to be missing. This does not appear to
affect IRB run normally.

This workaround forces RubyGems to eagerly load if it is enabled
and our lazy RubyGems support is active.