Fix interrupt handling in synchronize
The logic in MonitorMixin#mon_synchronize is broken as described
in ruby/monitor#2 due to the improper use of handle_interrupt.
This patch fixes the issue by ensuring the entire body of the
method is protected from interrupts except for the yield to the
incoming block. This avoids the ensure being interrupted, leaving
the monitor locked.

Fixes #6526