Prefer fcntl for checking fd liveness
The file descriptor created by inotify fails to fstat when running
inside the Windows Subsystem for Linux, resulting in an EPERM and
breaking JRuby on Rails apps that are using rb-inotify. This patch
mimics the CRuby code in preferring fcntl over fstat.

Since we cannot easily detect the availability of fcntl, this
logic just falls back to fstat for the one platform where we know
we can't use fcntl: Windows.

Fixes #6129.