Do not raise within rb_waitpid equivalent
The rb-waitpid function in CRuby does not actually trigger an
exception raise when the waitpid errors, leaving that to callers
to decide. Because ours did so, any errors during the
Process.detach call to waitpid would trigger the raise event hook
and unfix #6466.

This change moves the errno check outside of our rb_waitpid
equivalent, making it safe to use in the Process.detach thread.

Fixes #6466 again.