Unwrap delegated refined methods
Refinements installed over the top of an existing method get
readded to the method table with a RefinedWrapper dynamic method
wrapper. When this is combined with our Synchronized module, it
gets wrapped again with a SynchronizedDynamicMethod that the
refinment logic is unaware of. As a result, the eventual retrieval
of the inner wrapped method does not happen and methods that are
clearly defined trigger method_missing.

The fix here detects the delegated wrapper and unwraps it before
proceeding to the rest of the refinement logic, restoring the
proper behavior.

This is only a partial fix, since there are other places where the
wrapping can interfere with method lookup logic. It partially
addresses #6547 but a more complete fix is still needed.