Avoid write target modifying shared src bytes
When the incoming bytes are on the heap, we can use them in-place
for the string we pass to the Ruby write method. However, since
the target IO-like may want to modify the incoming string, we must
make sure it is marked as shared so our original src bytes will
not be modified.

This issue originally manifested in #4903 and was fixed by always
copying the incoming buffer, but marking the string shared has the
same effecit if the target attempts to make any modifications.