Fixes #6540. difference with block keyword arguments.
Our call logic was only using first incoming args[0] is arity was 1
and there is no kwrest(**a).  This should be arity of 1 and the lack of
 presence of any kwargs at all.

Intuitively, we know if we have kwargs at the callsite and the block
accepts kwargs...we should pass them forward.

We also know that procs will pass any extra args to the proc whether
the proc is defined to accept them or not.  As opposed to lambdas which
would error because you are passing the wrong number.

So I feel pretty good about this fix but I am very leery on landing this
the day before a release.  Blocks are tricky internally in Ruby.