Only increment count when adding a binding
The logic here increments the binding count to know whether we
have seen too many blocks for this yield optimization to help. In
the current logic, up to two blocks are allowed to bind directly
and all others will then be called along the indirect path.
However those other blocks will continue to increment the binding
count, eventually causing it to overlow the 32-bit signed int.
When that happens, the site will start eagerly binding every block
that comes in, an if enough are seen it will potentially blow the
stack and cause the yields to be very slow as they check hundreds
of different blocks for equality.

The fix here increments the binding count only along the path that
adds bindings. The non-binding path only reads the value.

Fixes #6260.