Mortally wound Arity to elevate Signature to the throne.
We have had a long-term goal to replace Arity with Signature (started in
2014?).  The main issue is that Arity had no concept of keyword arguments.
The math behind arity lives on via Ruby's arity method but the underlying
structure needs more information.

This commit completely deprecates all use of arity internally.  There
are a few removals of long deprecated code using Arity but we still have
a lot of deprecated code just in case.  For example, jruby-rack still
depends on this on one constructor still which receives Arity.

Anything which was using arity is now written in terms of Signature.  The
largest missing piece of DynamicMethod not having a getSignature.  As
mentioned in the last paragraph we only have a single live code path using
Arity from jruby-rack.

On the performance front we will do less work (although I doubt we will
see much we can measure).  In startings up Rails server on a single
controller app we were doing 3k arity calcs and over 400k signature calcs.
Now we are doing no arity calcs and only 2k signature calcs.  With some
better mechanism for Signature creation we could get this down to a few
hundred.  The second aspect of this is reducing memory about 1Mb for the
same Rails app.  Signature is about 100k now but since it is immutable we
can reduce this down to close to 0 if we share all combos during creation.

The more near term motivation was seeing a mixture of logic between arity
and signature involved in block argument processing.  This branch has already
removed some complexity of block processing but every extra detail we need
to keep in our head makes fixing block issues that much harder.