Try reducing n^2 complexity here
This should not time out, but if GC takes a while to process weak
references the original code will scan over 50M weakrefs and run
full GC 10k times before giving up. This change reduces the total
number of weakrefs to 1000 and only scans them with GC.start up
to 1000 times. If that is not enough GC to trigger weakref
evacuation then it should just fail.