Deprecate public APIs referencing Joda Time
Joda Time is being deprecated in favor of the Java 8 java.time API
and we need to start moving away from the former. This patch marks
all public APIs that accept or return Joda DateTime as deprecated
and provides java.time.ZonedDateTime versions for migration. This
patch does NOT convert any internals to using java.time APIs.

There are other Joda classes referenced in public APIs. so this is
only part of the work needed for #6600.

I have not suppressed deprecation warnings so we can see where we
still call the deprecated public APIs in JRuby.

This makes use of the MIT-licensed "timeywimey" converted class,
see https://github.com/meetup/timeywimey.