Bind compiled methods using lambdas
This change modifies CompiledIRMethod to use a lambda-implemented
interface rather than calling through MethodHandle when the target
method handles are provided with a Lookup that can see the
compiled method class. This improves stack utilization and call
performance.

This mechanism currently only works for "def" that occurs in
compiled bytecode, since the compiled class is not visible to the
LambdaMetaFactory otherwise. Jitted methods will need to provide
their Lookup when being bound into a MixedModeIRMethod in order to
duplicate this logic.

See #3810 for the initial motivation for this work.