Unbreak CachingCallSite.isBuiltin
During the various deprecations and undeprecations of this method,
it was inadvertently broken due to switching to a cacheAndGet path
that does not set the builtinCache field. Since call sites used
for isBuiltin are usually *only* used for isBuiltin, I modified
this to just use the standard cache field.

Relates to #6628