Improve inlining of varargs path for split methods
When we generate the invoker class for arity-split methods, we
only generate overrides of (DynamicMethod subtype) JavaMethod.call
signatures that align with the split signatures, including whether
or not they receive a block. This allows the base JavaMethod to
maintain a minimal set of base implementations, with unimplemented
arities and block paths falling back on default implementations.

Unfortunately due to our indy dispatch logic only being smart
enough to directly bind the lowest specific arity, this can break
inlining for all other arities if the block/no-block signatures
do not match, falling back on the base implementations in
JavaMethod.

This patch improves the "generic" DynamicMethod.call binding to
avoid passing a block when the target methods is backed by a
direct native handle and that handle wants a block. Because we
only allow homogeneous block/no-block signatures for arity-split
methods, omitting the block allows inlining straight through the
JavaMethod invoker even when calling the other arities.