Avoid using context backref for String#split
String#split appears to do nothing with the context backref slot
except using it to receive the MatchData from a regexp match and
eventually clearing it to nil. This patch sets up a thread-local
IRubyObject[] holder to use for that out value and avoids using
the context backref at all except for clearing it at the end.

This fixes #6160 by avoiding all reads of the context backref. It
may still cause problems for other backref-reading functions that
run concurrently and see the clearing to nil unexpectedly.