Directly call the block underlying Proc#call
This is a POC of a direct invocation path through Proc#call to the
block it eventually invokes. We use the call handle from a jitted
block as the endpoint and guard based on the identity of the proc
object. This limits the optimization to the first proc seen, so
this should not be seen as a complete solution.