Strict enforcement of MatchData usage
Modify setBackRef to want a MatchData so it can always set it as
in use. Clearing to nil should now be through clearBackRef. All
call paths updated to the new regime.

The justification here is that once a MatchData has been injected
into a frame, it is now visible to other threads. Whether it will
actually be read or not, we must say goodbye to that object since
it could be in use at any time on another thread that shares the
same frame.