Eliminate all reads of backref in core methods
This changes all places we read the frame-local backref to use the
thread-local match instead. None of these places used the backref
for its content; they only used it to reuse the object, which is
now handled more correctly by the thread-local match and the
setBackRef method that marks matches as in-use.

This should eliminate all internal data races against the $~
variable, since no core methods depend upon its value. This does
not make $~ thread-safe but it pushes the issue to the edge, where
the user will have to decide to use it knowing it may be updated
across threads.

See #4868 for the original issue that this partially fixes, and
see #6640 for an issue caused by these internal data races.