Sync the underlying OutputStream
Channel is intended to be an unbuffered IO construct backed
directly by a native file descriptor, but it has subclasses that
wrap other types of IO streams. In the case of OutputStream, the
wrapper seems to have a crucial flaw: it does not flush the
underlying stream. Since Channel provides no way to flush, the
buffered data will sit there forever.

This change modifies our logic to use a "sync" version of the
OutputStream channel wrapper, forcing the underlying stream to
flush after each write operation.