Unsupported process waits warn and noop
When running without native support, we cannot do any of the pid-
based waits. However in many instances it may be more damaging to
raise a hard error for the unimplemented logic rather than just
warning and doing nothing. This commit changes the logic of these
methods to report not implemented and warn rather than failing
outright. The methods are also marked as not implemented in case
code checks this (though I have no such examples).