Include overriding classes in backtrace mapping
Our Ruby backtrace-generation logic depends on a mapping of class
and method names to their equivalent Ruby name, but this fails
when the JRubyMethod annotation is on an abstract method
implemented by one or more subclasses. The subclasses are not
registered in the mapping and when their overridden methods appear
in the Java stack trace they will not be replaced with their Ruby
equivalent.

This patch adds support for JRubyMethod.implementers, a list of
additional classes to include in the mapping table. This is a
naive change currently, in that it assumes all methods on class A
should also be mapped for class B < A regardless of whether they
are actually overridden or not.

This is a first pass at fixing the missing trace lines reported in
issue #5200.