Allow argument forwarding to make it past parsing.
I have not examined how we handle the default values(&,*,**) but we
define them as lvars and will pass on '...'.  At this point we just
make it not choke on parsing to let it choke on building something
which has the wrong arity.

Current goal is to make parsing not die on any Ruby 3 syntactical
features.