Simplify mapping
The old logic was attempting to add all Ruby method names for a
given Java name but since we can only store one value the last
name would win. This modifies the logic to only attempt to add the
first name (or the simple name of the method, if no others are
specified) to the table.