Move implementers to class level
After some investigation it is clear that the "smart" approach of
finding all subclasses that override a given method is infeasible:

* Neither the Java compiler API nor the reflection API provide a
  way to walk all subclasses of a given class.
* Searching for overrides would trigger exceptions for each
  checked signature that is not overridden.

And since having the subclasses listed on each bound method would
be very fragile, this commit moves that list to the class level on
the JRubyClass annotation (if present) and simply duplicates the
parent's mapping list for each specified child. Children can still
have additional bindings that do not appear in the parent's set.