Redo case to be case and patterncase.
It looks like Ruby decided to not add an extra keyword and use 'case'
for both but they are not allowed to mix when and if.  This makes it
simpler just to create a new case called PatternCaseNode.

At this point nothing is processing the pattern case but it is wired
up to the point of implementing it within IRBuilder.  Getting some
partial or full implementation is the next step.