Support some pattern matching
Array pattern matching is working?  At this point I am still at the point
of just trying random samples vs a full test suite but the common stuff is
fine.  Here is an example:

```ruby
case [4, 2, 3, [5, 6]]
in [4, *, [5, a]]
  puts "ARR #{a}"
else
  puts "ELSE"
end
```

So pretty complicated and working.  The implementation is really what I
would call as roughed in.  The branch + jump strategy ended up backwards
where every success falls into a jump to the next test (in a series).  This
can be changed that a failure jumps (to next group).  Also the new instrs
and operands are worth reconsidering.  integer could potentially be
represented as fixnum and then the math op would work on longs vs ints.
At this point it is about making something which works and then going back
and making a second pass based on lessons learned in this one.

Next phase will be hash pattern.  After that point I will probably do more
exhaustive testing.