Golden path of hash pattern matching working.
Missing keyrest and some error conditions but basics are working:

```ruby
case {a: 1, b: 2}
in a: d
  puts "HASH #{d}"
else
  puts "ELSE"
end
```

At this point I might yak shave and try and make some of the basic patterns
of instrs used in pattern matching and frankly all over IRBuilder.  pattern
matching exposes how difficult it is to track labels and jumps when working.