Try something a bit different...friendly IRBuild methods.
Somewhat of an experiment which I am only using on pattern matching.
Use simpler method names and remove boiler plate.  The simpler names also
handle instruction adding to the scope.  This makes reading the build
code much easier and also more difficult.  The order these simpler
methods are called will also mean when the instructions are added.  This
coupling might make people accidentally use the simple method name in a
variable assignment not realizing they are ALSO adding the instr to the
scope.

This trade-off I am playing with.  At some level it is so much easier to
read what is happening that I feel the downside of the addInstr as a
side-effect will not be a large issue.  Also, as an issue, it is likely
to be confusing at the moment we add the code and not something we will
realize days or longer from when we added it.  Time will tell.