All pattern specs run on ruby/spec.
There is a missing experimental feature and I can see 15 or so issues on
MRI so we are definitely not done (also missing some JIT stuff for the
new instrs) but it is very close to be done.

One wrinkle is I added a dyn call to ! to reverse the value for unless
arms of patterns.  This should not be a dyn call.  I just put a FIXME for
now.  The alternative is to just add an extra label and branch which will
invert the value.  I am just moving forward for now and will circle back
to this.