Everything but argument handling logic for ruby2_keyword_hash.
There missing bit is marking the found keyword and handling the splat
differently.  This is an important part but this commit handles the
before and after of the behavior.

It is possible how this is set within both block and method will change
because we need to access it simply.  Any changes to what this commit
does should still be pretty small.