Implement Module#const_source_location.
This passes all specs from ruby/spec.  I have yet to try MRI internal
tests but the basics seem to be there.

There is an outstanding problem which may be large or small but I am punting
for now.  source location is depending on context.getFile() and
context.getLine() to return valid locations.  It does in the interpreter but
does not always in the bytecode compiler.  This may be due to some optimization
eliminating storage for file/line?  Did not look but this can be a second pass
activity.