Modify IO in and out streams to use sysread/write
This change addresses the remaining issue in #6588 by using
sysread and syswrite for the getInStream/getOutStream IO wrappers
that implement InputStream and OutputStream. This bypasses
encoding logic (as well as buffering), avoiding the question of
what encoding to treat incoming or outgoing bytes as.

I also added fast paths for sysread/syswrite that avoid allocating
any strings or byte arrays to accommodate this new logic.

Fixes #6588