Use java_home command on MacOS
When no JAVA_HOME env is set on MacOS, we do not properly detect
the actual JDK install location. The default `java` command is an
executable in /usr/bin/java that launches the default installed
JDK (usually the highest version). Since it is not a simlink the
logic we use on Linux to dig out the real JAVA_HOME does not work.

This patch uses the `java_home` command when no JAVA_HOME is set.
This command is only available on MacOS as far as I know, though
similar commands exist on other platforms.

Fixes #6726.

The same change needs to go into jruby-launcher.