Implement new case logic+.
The + is that I cleaned up notCompilable exceptions into a nice helper
which will now print out the location and the node (if provided) so it
is easier to see where something went wrong.

The new case logic itself will largely break about args{cat,push} and
listnodes (except not literal arrays) into more branches in IR.  This
fixes #6173 so that we evaluate less at the cost of having more branching.