Force build-time wrappers to run 'jruby'
On Windows, the default logic (from rubygems/defaults/jruby) will
see that we are running the build from within the 9.1.17.0 jar
file, and produce .bat wrappers that attempt to run JRuby using
a `java` command line. This breaks those .bat files, as described
in #6745, because the `java` command line gets completely wrapped
in quotes and no longer can be run as a single executable.

This patch un-patches the default Gem.ruby so that it goes back to
just using `jruby` as the command line to launch JRuby.

This is a pretty nasty hack only necessary because the build runs
the gem installation from within the 9.1.17.0 jar file. If the
installs were run from a normal JRuby command line (i.e. the
current working copy's bin/jruby) the wrappers are generated
correctly.