Only replace cmdline for 1:1 matching args
We parse out single-string commands along this path, which may
result in a single script with embedded spaces being broken into
individual args. Here we check that the original rawArgs length
matches the parsed args length so we know we will only replace the
argv[0] executable and not the entire script.