Only use cref class for autoload
The frame class is frequently pointing at the include wrapper for
a given module, in order to indicate where super should start
from. However with the "real module" logic this causes the target
of the autoload to not be on that included module, but on the
superclas of the include wrapper. This led to breakage in the
Kernel#autoload specs, since they require that Kernel.autoload and
Kernel#autoload define the constant on the included module.