Use cbase, not cref
We do not have a concept of cbase in our codebase, but the cbase
in CRuby is used as the target for class variables and other
class-y things that can be done from within a method body. The
logic in IRRuntimeHelpers.getModuleFromScope seems to
find the appropriate location.

We do not do class variable targeting exactly right (e.g. in
cloned modules) so this logic is not quite matching CRuby, but it
does pass the original case in #6708.

This still fails a regression test from "JRUBY-6570" but it
appears that the original fix for that issue did not really fix
the problem. It merely made the expected case pass by always using
Object as the target, which is clearly not correct.

In order to start using the language of CRuby I have introduced
a new method getCurrentClassBase that should eventually be
equivalent to CRuby's vm_get_cbase.