Remove questionable regression test to rubyspec
This spec defines current behavior of CRuby, but that behavior is
rather questionable due to unexpectedly different targeting of
autoload versus constant definition. We believe this regression
spec was never truly fixed in JRuby and merely happened to pass,
due to a naive fix that was still incorrect behavior but close
enough to make the given test pass. In order to move forward with
the fixes for #6708 we are moving this spec into rubyspec via
ruby/spec#839, so we can discuss whether this is really correct
behavior or not.