Reduce the hackery of the old fix of #1989.
lib/pom.rb was changing the permissions of all gem files to be 644.  This
caused other gems with binaries to lose their execute bit.  It also adds
time to the build during development.

The new code will only fix up the permissions of the unsupported gem
which demonstrates the issue.