Fixes #6242. Sorbet srb init crashes
The fix here is to not register JavaPackage as a constant of Java.
This prevents srb from finding it via constants.  So the problem class
is never seen.

Note: I think in the scenario of walking ObjectSpace we can still run
across this but that may be another issue to deal with when we experience it.

The explanation of what is wrong is that JavaPackage is in fact a Class but
due to how we want it to behave we set its metaclass to Module.  When srb
retrieves an UnboundMethod off of Class that method cannot be bound to this
type because it says "nope I am a module" (even though it is actually a class).