Trigger jit on return to avoid interp-after-jit
Because blocks are more sensitive to the surrounding scope, our
JIT will always interpret a block that has not yet been jit
compiled so that it can exit and be re-activated properly. This
caused an issue when using a //o regexp and jit threshold of zero,
because the jit happens before that first interpret, causing both
interpreted and jitted code to produce their own regexp object.

The fix here moves the jit request after the interpret, better
matching the expected behavior and guaranteeing the block will
finish interpreting at least once before the jit fires. As a
result, the jit sees the already-encountered //o regexp and can
compile it as a static regexp.

See discussion under #6753.