Fixes #6748. Inconsistent encoding when calling #to_s on arrays.
The main problem with the original code was that we would:

```java
    str.setEncoding(s.getEncoding());
```

Where s is an inspected element of the array.  MRI will call
rb_enc_associate which may setEncoding but might not.  In fact,
rb_enc_associate has a lot of logic which I don't really understand.
What I mean by that is that we call str.cat19 which will also do the
appropriate encoding negotiation of building up the result inspect
string.  Why add this extra method into the mix?

I opted to just let cat19 do its thing and spec:ruby:fast passed.

A second change I made was to make the inspect string use default_external
unless it is an encoding which does not support ascii (which is needed since
inspect is adding ascii like '[').  In the non-ascii case we just default to
US-ASCII.  Thankfully spec/ruby already has specs for this.