Lazily define FFI-using File methods
Activating FFI this early prevents it being upgraded by RubyGems
or Bundler. This also happens so early that there's no way to
eagerly activate the gem to pick up a newer version. As a result,
bundles or gems that depend on newer FFI will error or warn about
the mismatch.

This change modifies the FFI-based File methods on Windows
(File.symlink) and Solaris (File#flock) to lazily require a
separate file that loads FFI and redefines those methods.

Fixes #6751